Copy XYPlot to Clipboard.
