+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/menu/DebriefRibbonFile.java
@@ -1,16 +1,8 @@
 package org.mwc.debrief.lite.menu;
 
 import java.awt.Dimension;
-import java.awt.Graphics;
 import java.awt.Image;
-import java.awt.Toolkit;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.ClipboardOwner;
-import java.awt.datatransfer.DataFlavor;
-import java.awt.datatransfer.Transferable;
-import java.awt.datatransfer.UnsupportedFlavorException;
 import java.awt.event.ActionEvent;
-import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -69,59 +61,7 @@ public class DebriefRibbonFile
     public void actionPerformed(final ActionEvent e)
     {
       final JMapPane map = (JMapPane) mapRenderer.getMap();
-      final BufferedImage img = new BufferedImage(map.getWidth(), map
-          .getHeight(), BufferedImage.TYPE_INT_ARGB);
-      final Graphics g = img.getGraphics();
-      map.paint(g);
-      g.dispose();
-
-      if (img != null)
-      {
-        final Transferable t = new Transferable()
-        {
-
-          @Override
-          public Object getTransferData(final DataFlavor flavor)
-              throws UnsupportedFlavorException, IOException
-          {
-            if (isDataFlavorSupported(flavor))
-            {
-              return img;
-            }
-            return null;
-          }
-
-          @Override
-          public DataFlavor[] getTransferDataFlavors()
-          {
-            return new DataFlavor[]
-            {DataFlavor.imageFlavor};
-          }
-
-          @Override
-          public boolean isDataFlavorSupported(final DataFlavor flavor)
-          {
-            if (flavor == DataFlavor.imageFlavor)
-              return true;
-            return false;
-          }
-
-        };
-
-        final ClipboardOwner co = new ClipboardOwner()
-        {
-
-          @Override
-          public void lostOwnership(final Clipboard clipboard,
-              final Transferable contents)
-          {
-            System.out.println("Copy to PNG: Lost Ownership");
-          }
-
-        };
-        final Clipboard cb = Toolkit.getDefaultToolkit().getSystemClipboard();
-        cb.setContents(t, co);
-      }
+      org.mwc.debrief.lite.util.ClipboardUtils.copyToClipboard(map);
     }
   }
